fix: passport compressor 에러 핸들링
